Start moving into new VirtualItem
