+++ reactions/src/main/java/me/fromgate/reactions/data/ItemStackValue.java
@@ -1,7 +1,7 @@
 package me.fromgate.reactions.data;
 
 import me.fromgate.reactions.util.item.ItemUtils;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import me.fromgate.reactions.util.location.LocationUtils;
 import org.bukkit.Location;
 import org.bukkit.inventory.ItemStack;
@@ -40,7 +40,7 @@ public class ItemStackValue implements DataValue {
 
     @Override
     public boolean set(String value) {
-        ItemStack item = VirtualItem.fromString(value);
+        ItemStack item = LegacyVirtualItem.fromString(value);
         if (item != null) {
             this.value = item;
             return true;
