+++ reactions/src/main/java/me/fromgate/reactions/menu/InventoryMenu.java
@@ -4,7 +4,7 @@ import me.fromgate.reactions.ReActions;
 import me.fromgate.reactions.module.basics.StoragesManager;
 import me.fromgate.reactions.util.FileUtils;
 import me.fromgate.reactions.util.item.ItemUtils;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import me.fromgate.reactions.util.message.Msg;
 import me.fromgate.reactions.util.parameter.Parameters;
 import org.bukkit.Bukkit;
@@ -128,7 +128,7 @@ public class InventoryMenu implements Listener {
             for (int i = 1; i <= size; i++) {
                 String slotStr = "slot" + i;
                 if (!param.contains(slotStr)) continue;
-                ItemStack slotItem = VirtualItem.fromString(param.getString(slotStr));
+                ItemStack slotItem = LegacyVirtualItem.fromString(param.getString(slotStr));
                 if (slotItem == null) continue;
                 inv.setItem(i - 1, slotItem);
             }
@@ -202,7 +202,7 @@ public class InventoryMenu implements Listener {
 
     private static String itemToString(String itemStr) {
         if (itemStr.isEmpty()) return "AIR";
-        ItemStack item = VirtualItem.fromString(itemStr);
+        ItemStack item = LegacyVirtualItem.fromString(itemStr);
         if (item == null || item.getType() == Material.AIR) return "AIR";
         String returnStr = item.hasItemMeta() && item.getItemMeta().hasDisplayName() ? item.getItemMeta().getDisplayName() : "";
         String itemTypeData = item.getType().name() + (ItemUtils.getDurability(item) == 0 ? "" : ":" + ItemUtils.getDurability(item)) + (item.getAmount() == 1 ? "" : "*" + item.getAmount());
