+++ reactions/src/main/java/me/fromgate/reactions/module/basics/actions/ActionBlockFill.java
@@ -27,7 +27,7 @@ import me.fromgate.reactions.logic.RaContext;
 import me.fromgate.reactions.logic.activity.actions.Action;
 import me.fromgate.reactions.util.Rng;
 import me.fromgate.reactions.util.alias.Aliases;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import me.fromgate.reactions.util.location.LocationUtils;
 import me.fromgate.reactions.util.message.Msg;
 import me.fromgate.reactions.util.parameter.Parameters;
@@ -50,7 +50,7 @@ public class ActionBlockFill implements Action {
         Parameters itemParam = Parameters.fromString(params.getString("block", "AIR"), "type");
         ItemStack item = null;
         if (!itemParam.getString("type", "AIR").equalsIgnoreCase("air")) {
-            item = VirtualItem.fromMap(itemParam.getMap());
+            item = LegacyVirtualItem.fromMap(itemParam.getMap());
             if (item == null || !item.getType().isBlock()) {
                 Msg.logOnce("wrongblockfill" + params.getString("block"),
                         "Failed to execute action BLOCK_FILL. Wrong block " + params.getString("block"));
