+++ reactions/src/main/java/me/fromgate/reactions/module/basics/activators/ItemClickActivator.java
@@ -29,7 +29,7 @@ import me.fromgate.reactions.module.basics.storages.ItemClickStorage;
 import me.fromgate.reactions.util.Utils;
 import me.fromgate.reactions.util.enums.HandType;
 import me.fromgate.reactions.util.item.ItemUtils;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import me.fromgate.reactions.util.message.Msg;
 import me.fromgate.reactions.util.parameter.Parameters;
 import org.bukkit.configuration.ConfigurationSection;
@@ -59,7 +59,7 @@ public class ItemClickActivator extends Activator {
 
     @Override
     public boolean checkStorage(Storage event) {
-        if (item.isEmpty() || (VirtualItem.fromString(item) == null)) {
+        if (item.isEmpty() || (LegacyVirtualItem.fromString(item) == null)) {
             Msg.logOnce(logic.getName() + "activatoritemempty", "Failed to parse item of activator " + logic.getName());
             return false;
         }
