+++ reactions/src/main/java/me/fromgate/reactions/module/basics/activators/ItemHoldActivator.java
@@ -29,7 +29,7 @@ import me.fromgate.reactions.module.basics.storages.ItemHoldStorage;
 import me.fromgate.reactions.util.Utils;
 import me.fromgate.reactions.util.enums.HandType;
 import me.fromgate.reactions.util.item.ItemUtils;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import me.fromgate.reactions.util.message.Msg;
 import me.fromgate.reactions.util.parameter.Parameters;
 import org.bukkit.configuration.ConfigurationSection;
@@ -63,7 +63,7 @@ public class ItemHoldActivator extends Activator {
 
     @Override
     public boolean checkStorage(Storage event) {
-        if (itemStr.isEmpty() || (VirtualItem.fromString(itemStr) == null)) {
+        if (itemStr.isEmpty() || (LegacyVirtualItem.fromString(itemStr) == null)) {
             Msg.logOnce(logic.getName() + "activatorholdempty", "Failed to parse itemStr of activator " + logic.getName());
             return false;
         }
