+++ reactions/src/main/java/me/fromgate/reactions/module/basics/storages/ItemClickStorage.java
@@ -28,7 +28,7 @@ import me.fromgate.reactions.logic.activators.Activator;
 import me.fromgate.reactions.logic.activators.Storage;
 import me.fromgate.reactions.module.basics.activators.ItemClickActivator;
 import me.fromgate.reactions.util.collections.MapBuilder;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
@@ -55,7 +55,7 @@ public class ItemClickStorage extends Storage {
     @Override
     protected @NotNull Map<String, String> prepareVariables() {
         Map<String, String> tempVars = new HashMap<>();
-        VirtualItem vItem = VirtualItem.fromItemStack(item);
+        LegacyVirtualItem vItem = LegacyVirtualItem.fromItemStack(item);
         if (item != null) {
             tempVars.put("item", vItem.toString());
             tempVars.put("item-str", vItem.toDisplayString());
