+++ reactions/src/main/java/me/fromgate/reactions/module/basics/storages/ItemHeldStorage.java
@@ -6,7 +6,7 @@ import me.fromgate.reactions.logic.activators.Activator;
 import me.fromgate.reactions.logic.activators.Storage;
 import me.fromgate.reactions.module.basics.activators.ItemHeldActivator;
 import me.fromgate.reactions.util.collections.MapBuilder;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
@@ -41,14 +41,14 @@ public class ItemHeldStorage extends Storage {
     protected @NotNull Map<String, String> prepareVariables() {
         Map<String, String> tempVars = new HashMap<>();
         if (newItem != null) {
-            VirtualItem vi = VirtualItem.fromItemStack(newItem);
+            LegacyVirtualItem vi = LegacyVirtualItem.fromItemStack(newItem);
             if (vi != null) {
                 tempVars.put("itemnew", vi.toString());
                 tempVars.put("itemnew-str", vi.toDisplayString());
             }
         }
         if (previousItem != null) {
-            VirtualItem vi = VirtualItem.fromItemStack(previousItem);
+            LegacyVirtualItem vi = LegacyVirtualItem.fromItemStack(previousItem);
             if (vi != null) {
                 tempVars.put("itemprev", vi.toString());
                 tempVars.put("itemprev-str", vi.toDisplayString());
