+++ reactions/src/main/java/me/fromgate/reactions/module/basics/storages/ItemHoldStorage.java
@@ -25,7 +25,7 @@ package me.fromgate.reactions.module.basics.storages;
 import me.fromgate.reactions.logic.activators.Activator;
 import me.fromgate.reactions.logic.activators.Storage;
 import me.fromgate.reactions.module.basics.activators.ItemHoldActivator;
-import me.fromgate.reactions.util.item.VirtualItem;
+import me.fromgate.reactions.util.item.LegacyVirtualItem;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
@@ -52,7 +52,7 @@ public class ItemHoldStorage extends Storage {
     @Override
     protected @NotNull Map<String, String> prepareVariables() {
         Map<String, String> tempVars = new HashMap<>();
-        VirtualItem vItem = VirtualItem.fromItemStack(item);
+        LegacyVirtualItem vItem = LegacyVirtualItem.fromItemStack(item);
         if (item != null) {
             tempVars.put("item", vItem.toString());
             tempVars.put("item-str", vItem.toDisplayString());
